so many changes...
for flow, fixed a lot of problems w.r.t. bascheck.. and the handling of
trajectories that are cut off due to 'count'.

put in read-modify-write's for everything that needs them, and made
extra exclusion rules for the doubles to make sure everything is kosher.

writes out 'fprob.dat', and I'm trying to get all of the principal
crossings at least once so I can use the matrix-based alg.